GST:Added Repository and DB connection

+++ src/main/java/com/gst/gst_calculator/repository/ItemRepository.java
@@ -6,6 +6,8 @@ import com.gst.gst_calculator.models.Item;
 
 
 @Repository
-public interface ItemRepository extends JpaRepository<Item, Long> {
+public interface ItemRepository extends JpaRepository<Item, Integer> {
+
+    
     
 }
